Update ArubaClearpassAPI
This script first sets the API endpoint URL and the credentials of a user with access to the API. It then uses curl to send a POST request to the oauth endpoint to obtain an access token using the credentials.

The response from the API contains a JSON object that includes the access token. The script extracts the token using the jq command and stores it in the ACCESS_TOKEN variable.

Finally, the script uses the access token in an HTTP GET request to the endpoint API endpoint and stores the response in the API_RESPONSE variable. The script then displays the response.

Note that this script assumes you have jq installed to parse JSON responses from the API. If you don't have jq installed, you can install it on most Linux distributions using your package manager.